requirements.txt
vega-datasets
streamlit
obspy
PyWavelets
vega-datasets
obspy
altair
scikit-learn
panda
numpy